fix: Set file extensions for package entry files
If the package type is CommonJS all `.js` files will be handled
as if they are CommonJS, even if exports is set to module.
Same if type is set to ESM, then all `.js` files are handled as
if they are modules. So the file extension has to be set to the explicit type.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>