chore: adjust `engines` versions to support wider range
This library is used by multiple apps also supporting wide range of
Nextcloud versions, thus we need to support the current Node version
(22) as well as the one used by the oldest maintained Nextcloud version
(Nextcloud 30 = Node 20).
Future proof we also support Node 24 - as this package is a browser
library and not a NodeJS library its fine!

Move package manager to `devEngines` as its not a runtime requirement.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>